Refactor dependency injection
This commit refactors the dependency injection setup in the application. It introduces a new `CoreDI` class to manage core dependencies like `DioClient` and `ApiService`.
The `DependencyInjector` is updated to use `CoreDI` and feature-specific dependency injection classes such as `AuthDI`.
The main application entry point is modified to utilize the new dependency injection setup.

Key changes:

-   Created `core_di.dart` for managing core dependencies.
-   Updated `di.dart` to use `CoreDI` and feature-specific DI classes.
-   Updated `main.dart` to use the new `DependencyInjector`.
-   Removed unused `_runApp()` function in `main.dart`
- Added CacheService for DioClient.