Refactor: Product detail screen and gallery
This commit refactors the product detail screen to improve the display of product information. It removes the tab bar and instead displays the description and specifications directly on the screen. Additionally, it enhances the product gallery by integrating a fullscreen gallery view that can be opened by tapping on the images.